Use Strategies
How things currently are
---
Our strategy ideas (what to buy, what choices to make in what order,
etc) are all on Player. By separating this logic out into a separate
class we can have agents which are only concerned with what the player
*should* do, and don't have to implement any logic around what the
player *could* do. These objects essentially get:

- a game state
- a bunch of choices

From there they can return any of the choices they were given, and the
rest of the game will take care of carrying out the impacts of that
choice. This required some significant code changes because (at least
for me to implement this) it required a differentiation between the
*ability* to do a *category* of thing (for instance, all players always
have the buy, play, and attack abilities assuming there are valid
targets), but the buy ability is different from the following Choices:

- Buy Viper
- Buy Frigate
- Buy Imperial Whoseewhatsit
- Buy FlooberGunner
- Buy Trading Post

Those are the choices that might be generated by the combination of a
Buy ability and a particular trade row (and a particular amount of trade
held by the player). This becomes particularly clear when a player
chooses to use a buy ability, and now their buy choices might change!
(because the cards in the trade row have changed or they lost money and
now they can't afford as much). Abilities and choices are different
things, and the only 100% accurate way to produce a list of choices is
by iterating over the current abilities and expanding them with valid
targets to form choices.

This change is probably 90% of the actual work done here.